Stash fix
