+++ pages/4_📊_Correlations.py
@@ -12,7 +12,7 @@ from pingouin import rm_corr
 
 from datawizard.utils import trim_code, tuple2str
 from globals import MAX_VARIABLES_PLOT
-from st_widgets.commons import app_config, load_stash
+from st_widgets.commons import app_config, load_stash, read_stash_from_history
 from st_widgets.console import session_console
 from st_widgets.dataframe import empty_eurostat_dataframe
 from st_widgets.stateful.number_input import stateful_number_input
@@ -87,7 +87,7 @@ if __name__ == "__main__":
     try:
         with st.spinner(text="Fetching data"):
             if "history" in st.session_state:
-                stash = load_stash(st.session_state.history)
+                stash = load_stash(read_stash_from_history(st.session_state.history))
             else:
                 st.warning("No stash found. Select some data to plot.")
     except ValueError as ve:
